requirements.txt
pyinquirer==1.0.3
gitpython==2.1.11
termcolor==1.1.0
progressbar==2.5